Date helpers are now helper functions and not components anymore - it is quite confusing to use components for working with strings
